refactor: move constant out to avoid hook deps
