+++ src/pages/HardFork/comments.ts
@@ -1,0 +1,101 @@
+const comments = [
+  {
+    name: 'Nervos CKB Docs',
+    avatars: 'https://docs.nervos.org/img/logo-dark.png',
+    content: 'hardfork.comment1',
+    date: '',
+    link: 'https://docs.nervos.org/docs/history-and-hard-forks/intro-to-hard-fork',
+  },
+  {
+    name: 'Nervos CKB Docs',
+    avatars: 'https://docs.nervos.org/img/logo-dark.png',
+    content: 'hardfork.comment2',
+    date: '',
+    link: 'https://docs.nervos.org/docs/history-and-hard-forks/intro-to-hard-fork',
+  },
+  {
+    name: 'Nervos CKB Docs',
+    avatars: 'https://docs.nervos.org/img/logo-dark.png',
+    content: 'hardfork.comment3',
+    date: '',
+    link: 'https://docs.nervos.org/docs/history-and-hard-forks/intro-to-hard-fork',
+  },
+  {
+    name: 'doitian',
+    avatars: 'https://avatars.githubusercontent.com/u/35768?v=4',
+    content: 'hardfork.comment4',
+    date: '',
+    link: 'https://github.com/nervosnetwork/ckb/issues/4806#issuecomment-2673768017',
+  },
+  {
+    name: 'doitian',
+    avatars: 'https://avatars.githubusercontent.com/u/35768?v=4',
+    content: 'hardfork.comment5',
+    date: '',
+    link: 'https://github.com/nervosnetwork/ckb/issues/4806#issuecomment-2673786997',
+  },
+  {
+    name: 'doitian',
+    avatars: 'https://avatars.githubusercontent.com/u/35768?v=4',
+    content: 'hardfork.comment6',
+    date: '',
+    link: 'https://github.com/nervosnetwork/ckb/issues/4806#issuecomment-2673786997',
+  },
+  {
+    name: 'gpBlockchain',
+    avatars: 'https://avatars.githubusercontent.com/u/32102187?v=4&size=80',
+    content: 'hardfork.comment7',
+    date: '',
+    link: 'https://github.com/nervosnetwork/ckb/issues/4806#issuecomment-2677703878',
+  },
+  {
+    name: 'CKB Consensus Change',
+    avatars: 'https://avatars.githubusercontent.com/u/35361817?s=200&v=4',
+    content: 'hardfork.comment8',
+    date: '',
+    link: 'https://github.com/nervosnetwork/rfcs/blob/3a6ae4fa5d59b6e33fa7bd563d336706d135c0d8/rfcs/0053-ckb-hardfork/0053-ckb-hardfork.md',
+  },
+  {
+    name: 'CKB Consensus Change',
+    avatars: 'https://avatars.githubusercontent.com/u/35361817?s=200&v=4',
+    content: 'hardfork.comment9',
+    date: '',
+    link: 'https://github.com/nervosnetwork/rfcs/blob/3a6ae4fa5d59b6e33fa7bd563d336706d135c0d8/rfcs/0053-ckb-hardfork/0053-ckb-hardfork.md',
+  },
+  {
+    name: 'CKB Consensus Change',
+    avatars: 'https://avatars.githubusercontent.com/u/35361817?s=200&v=4',
+    content: 'hardfork.comment10',
+    date: '',
+    link: 'https://github.com/nervosnetwork/rfcs/blob/3a6ae4fa5d59b6e33fa7bd563d336706d135c0d8/rfcs/0053-ckb-hardfork/0053-ckb-hardfork.md',
+  },
+  {
+    name: 'CKB Consensus Change',
+    avatars: 'https://avatars.githubusercontent.com/u/35361817?s=200&v=4',
+    content: 'hardfork.comment11',
+    date: '',
+    link: 'https://github.com/nervosnetwork/rfcs/blob/3a6ae4fa5d59b6e33fa7bd563d336706d135c0d8/rfcs/0053-ckb-hardfork/0053-ckb-hardfork.md',
+  },
+  {
+    name: 'CKB Consensus Change',
+    avatars: 'https://avatars.githubusercontent.com/u/35361817?s=200&v=4',
+    content: 'hardfork.comment12',
+    date: '',
+    link: 'https://github.com/nervosnetwork/rfcs/blob/3a6ae4fa5d59b6e33fa7bd563d336706d135c0d8/rfcs/0053-ckb-hardfork/0053-ckb-hardfork.md',
+  },
+  {
+    name: 'CKB Consensus Change',
+    avatars: 'https://avatars.githubusercontent.com/u/35361817?s=200&v=4',
+    content: 'hardfork.comment13',
+    date: '',
+    link: 'https://github.com/nervosnetwork/rfcs/blob/3a6ae4fa5d59b6e33fa7bd563d336706d135c0d8/rfcs/0053-ckb-hardfork/0053-ckb-hardfork.md',
+  },
+  {
+    name: 'CKB Consensus Change',
+    avatars: 'https://avatars.githubusercontent.com/u/35361817?s=200&v=4',
+    content: 'hardfork.comment14',
+    date: '',
+    link: 'https://github.com/nervosnetwork/rfcs/blob/3a6ae4fa5d59b6e33fa7bd563d336706d135c0d8/rfcs/0053-ckb-hardfork/0053-ckb-hardfork.md',
+  },
+]
+export default comments
